@@ -5,6 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "tsc": "tsc",
     "clean": "rimraf www/ || true",
     "build": "npm run clean && tsc && cp package.json www/package.json && mkdir www/tmp/ && cd www && zip -r Archive.zip . && cd ..",
     "dev": "ts-node-dev --respawn --transpile-only ./src/server.ts"
@@ -23,9 +24,9 @@
     "@types/bluebird": "^3.5.33",
     "@types/express": "^4.17.0",
     "@types/node": "^11.13.17",
-    "ts-node-dev": "^1.0.0-pre.40",
-    "tslint": "^5.18.0",
-    "typescript": "^3.5.3"
+    "ts-node-dev": "^1.1.1",
+    "tslint": "^5.20.1",
+    "typescript": "^3.9.7"
   },
   "dependencies": {
     "express": "^4.17.1",
