@@ -1,5 +1,6 @@
 import fs from "fs";
-import Jimp = require("jimp");
+// import Jimp = require("jimp");
+import Jimp from 'jimp';
 
 // filterImageFromURL
 // helper function to download, filter, and save the filtered image locally
